requirements.txt
asgiref==3.9.1
Django==4.2.23
djangorestframework==3.16.1
PyMySQL==1.1.2
sqlparse==0.5.3
typing_extensions==4.15.0
